update-migration

+++ database/migrations/2024_08_24_111907_add_nominal_tagihan_to_history_table.php
@@ -11,10 +11,13 @@ class AddNominalTagihanToHistoryTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::table('history', function (Blueprint $table) {
-            $table->decimal('nominal_tagihan', 10, 2)->nullable(); // Menambahkan kolom nominal_tagihan
+            // Cek apakah kolom 'nominal_tagihan' sudah ada sebelum menambahkannya
+            if (!Schema::hasColumn('history', 'nominal_tagihan')) {
+                $table->decimal('nominal_tagihan', 10, 2)->nullable(); // Menambahkan kolom nominal_tagihan
+            }
         });
     }
 
@@ -23,10 +26,14 @@ class AddNominalTagihanToHistoryTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::table('history', function (Blueprint $table) {
-            $table->dropColumn('nominal_tagihan'); // Menghapus kolom nominal_tagihan jika rollback
+            // Cek apakah kolom 'nominal_tagihan' ada sebelum menghapusnya
+            if (Schema::hasColumn('history', 'nominal_tagihan')) {
+                $table->dropColumn('nominal_tagihan'); // Menghapus kolom nominal_tagihan jika rollback
+            }
         });
     }
 }
+
